+++ MultiSource/Olden/health/health.c
@@ -99,8 +99,8 @@ struct Results get_results(struct Village *village)
   list = village->returned.forward;
   while (list != NULL) {
     p = list->patient;
-    r1.total_hosps += (float)(p->hosps_visited);
-    r1.total_time += (float)(p->time); 
+    r1.total_hosps += (double)(p->hosps_visited);
+    r1.total_time += (double)(p->time); 
     r1.total_patients += 1.0;
     list = list->forward; }            /* :) adt_pf detected */
   
@@ -131,10 +131,10 @@ void check_patients_inside(struct Village *village, struct List *list)
 
 struct List *check_patients_assess(struct Village *village, struct List *list) 
 {
-  float                 rand;
+  double                 rand;
   struct Patient          *p;
   struct List             *up = NULL;
-  long                     s;
+  long long                     s;
   int                      label, t;
 
   while (list != NULL) 
@@ -146,7 +146,7 @@ struct List *check_patients_assess(struct Village *village, struct List *list)
       if (p->time_left == 0) { 
         s = village->seed;
         rand = my_rand(s);
-        village->seed = (long)(rand * IM);
+        village->seed = (long long)(rand * IM);
         label = village->label;
         if (rand > 0.1 || label == 0) {
           removeList(&village->hosp.assess, p);
@@ -213,14 +213,14 @@ void put_in_hosp(struct Hosp *hosp, struct Patient *patient)
 
 struct Patient *generate_patient(struct Village *village) 
 {
-  long            s,newseed; 
+  long long            s,newseed; 
   struct Patient *patient;
-  float rand;
+  double rand;
   int label;
   
   s = village->seed;
   rand = my_rand(s);
-  village->seed = (long)(rand * IM);
+  village->seed = (long long)(rand * IM);
   newseed = village->seed;
   label = village->label;
   if (rand > 0.666) {
@@ -239,7 +239,7 @@ int main(int argc, char *argv[])
   struct Results         results;
   struct Village         *top = 0;
   int                    i;
-  float                  total_time,
+  double                  total_time,
                          total_patients,
                          total_hosps;
